# backend/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Zdravdom API
  description: Home healthcare platform API
  version: 1.0.0
  contact:
    name: Zdravdom Development Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookingStatus:
      type: string
      enum: [REQUESTED, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, PAID, REFUNDED, FAILED]
...
    Platform:
      type: string
      enum: [IOS, ANDROID]
    ServiceCategory:
      type: string
      enum: [NURSING, THERAPY, LABORATORY, SPECIALIST]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request data
        field:
          type: string
          example: email
        timestamp:
          type: string
          format: date-time
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Home Nursing Care
        category:
          $ref: '#/components/schemas/ServiceCategory'
        description:
          type: string
        duration:
          type: integer
          description: Estimated duration in minutes
          example: 60
        price:
          type: number
          format: float
          example: 45.00
        currency:
          type: string
          default: EUR
        rating:
          type: number
          format: float
        reviewsCount:
          type: integer
        imageUrl:
          type: string
        includedItems:
          type: array
          items:
            type: string
    ServicePackage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        name:
          type: string
          example: S
        description:
          type: string
        price:
          type: number
          format: float
        duration:
          type: integer
          description: Duration in minutes
...
    TimeSlot:
      type: object
      properties:
        time:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "10:00"
        available:
          type: boolean
...
    CreateBookingRequest:
      type: object
      required:
        - serviceId
        - addressId
        - date
        - timeSlot
      properties:
        serviceId:
          type: string
          format: uuid
        packageId:
          type: string
          format: uuid
        addressId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        timeSlot:
          type: string
          example: "09:00"
        providerId:
          type: string
          format: uuid
          description: Optional preferred provider
        notes:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        packageId:
          type: string
          format: uuid
        addressId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        timeSlot:
          type: string
          example: "09:00"
        status:
          $ref: '#/components/schemas/BookingStatus'
        paymentAmount:
          type: number
          format: float
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        statusTimeline:
          type: array
          items:
            $ref: '#/components/schemas/StatusTimelineItem'
    StatusTimelineItem:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/BookingStatus'
        timestamp:
          type: string
          format: date-time
        note:
          type: string
...
    CancelBookingRequest:
      type: object
      properties:
        reason:
          type: string
    RejectBookingRequest:
      type: object
      properties:
        reason:
          type: string
    BookingListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ─── Visit ──────────────────────────────────────────────────────────────────
...
    RegisterPushTokenRequest:
      type: object
      required:
        - token
        - platform
      properties:
        token:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        data:
          type: object
          additionalProperties: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── Ratings ───────────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /services:
    get:
      summary: List all services (filter by category, search)
      tags: [Services]
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ServiceCategory'
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  page:
                    type: integer
                  size:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{id}:
    get:
      summary: Get service details
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{id}/packages:
    get:
      summary: Get S/M/L packages for a service
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicePackage'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Providers ────────────────────────────────────────────────────────────────
...
  /booking/slots:
    get:
      summary: Get available time slots for a service at an address
      tags: [Booking]
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: addressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Booking CRUD ─────────────────────────────────────────────────────────────
  /booking:
    post:
      summary: Create a new booking
      tags: [Booking]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Slot no longer available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/{id}:
    get:
      summary: Get booking details
      tags: [Booking]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/{id}/cancel:
    put:
      summary: Cancel a booking
      tags: [Booking]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Cannot cancel in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/{id}/accept:
    put:
      summary: Provider accepts a booking
      tags: [Booking]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Cannot accept in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/{id}/reject:
    put:
      summary: Provider rejects a booking
      tags: [Booking]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectBookingRequest'
      responses:
        '200':
          description: Booking rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Cannot reject in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/history:
    get:
      summary: Paginated booking history
      tags: [Booking]
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated booking history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /booking/upcoming:
    get:
      summary: Get active/upcoming bookings
      tags: [Booking]
      responses:
        '200':
          description: List of upcoming bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Visits ───────────────────────────────────────────────────────────────────
...
  /notifications/push-token:
    post:
      summary: Register device push token
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPushTokenRequest'
      responses:
        '201':
          description: Token registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token registered
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notifications:
    get:
      summary: List notifications for current user
      tags: [Notifications]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  page:
                    type: integer
                  size:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notifications/{id}/read:
    put:
      summary: Mark notification as read
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ─── Tags ─────────────────────────────────────────────────────────────────────
...