# backend/src/main/java/com/zdravdom/booking/domain/TimeSlot.java
package com.zdravdom.booking.domain;

import java.time.LocalTime;

/**
 * Time slot representing a booking window.
 */
public record TimeSlot(
    LocalTime startTime,
    LocalTime endTime
) {
    public TimeSlot {
        if (startTime == null) {
            throw new IllegalArgumentException("Start time cannot be null");
        }
        if (endTime == null) {
            throw new IllegalArgumentException("End time cannot be null");
        }
        if (startTime.isAfter(endTime) || startTime.equals(endTime)) {
            throw new IllegalArgumentException("Start time must be before end time");
        }
    }

    public int getDurationMinutes() {
        return (int) java.time.Duration.between(startTime, endTime).toMinutes();
    }

    public boolean overlaps(TimeSlot other) {
        return this.startTime.isBefore(other.endTime) && this.endTime.isAfter(other.startTime);
    }
}